Move Hibernate settings to a dedicated class
Closes gh-1327
