+++ spring-boot-project/spring-boot-autoconfigure/src/test/java/org/springframework/boot/autoconfigure/orm/jpa/CustomHibernateJpaAutoConfigurationTests.java
@@ -67,11 +67,13 @@ public class CustomHibernateJpaAutoConfigurationTests {
 				"spring.jpa.properties.hibernate.ejb.naming_strategy_delegator:"
 						+ "org.hibernate.cfg.naming.ImprovedNamingStrategyDelegator")
 				.run((context) -> {
-					JpaProperties bean = context.getBean(JpaProperties.class);
-					Map<String, Object> hibernateProperties = bean
-							.getHibernateProperties(new HibernateSettings());
-					assertThat(hibernateProperties.get("hibernate.ejb.naming_strategy"))
-							.isNull();
+					JpaProperties jpaProperties = context.getBean(JpaProperties.class);
+					HibernateProperties hibernateProperties = context
+							.getBean(HibernateProperties.class);
+					Map<String, Object> properties = hibernateProperties
+							.determineHibernateProperties(jpaProperties.getProperties(),
+									new HibernateSettings());
+					assertThat(properties.get("hibernate.ejb.naming_strategy")).isNull();
 				});
 	}
