+++ spring-boot-project/spring-boot-autoconfigure/src/test/java/org/springframework/boot/autoconfigure/orm/jpa/JpaPropertiesTests.java
@@ -19,21 +19,13 @@ package org.springframework.boot.autoconfigure.orm.jpa;
 import java.sql.Connection;
 import java.sql.DatabaseMetaData;
 import java.sql.SQLException;
-import java.util.Map;
 import java.util.function.Consumer;
-import java.util.function.Supplier;
 
 import javax.sql.DataSource;
 
-import org.hibernate.cfg.AvailableSettings;
-import org.junit.Before;
 import org.junit.Test;
-import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
 
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.boot.orm.jpa.hibernate.SpringImplicitNamingStrategy;
-import org.springframework.boot.orm.jpa.hibernate.SpringPhysicalNamingStrategy;
 import org.springframework.boot.test.context.assertj.AssertableApplicationContext;
 import org.springframework.boot.test.context.runner.ApplicationContextRunner;
 import org.springframework.boot.test.context.runner.ContextConsumer;
@@ -41,7 +33,6 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.orm.jpa.vendor.Database;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.entry;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
@@ -56,90 +47,6 @@ public class JpaPropertiesTests {
 
 	private final ApplicationContextRunner contextRunner = new ApplicationContextRunner()
 			.withUserConfiguration(TestConfiguration.class);
-
-	@Mock
-	private Supplier<String> ddlAutoSupplier;
-
-	@Before
-	public void setup() {
-		MockitoAnnotations.initMocks(this);
-	}
-
-	@Test
-	public void noCustomNamingStrategy() {
-		this.contextRunner.run(assertJpaProperties((properties) -> {
-			Map<String, Object> hibernateProperties = properties
-					.getHibernateProperties(new HibernateSettings());
-			assertThat(hibernateProperties)
-					.doesNotContainKeys("hibernate.ejb.naming_strategy");
-			assertThat(hibernateProperties).containsEntry(
-					"hibernate.physical_naming_strategy",
-					SpringPhysicalNamingStrategy.class.getName());
-			assertThat(hibernateProperties).containsEntry(
-					"hibernate.implicit_naming_strategy",
-					SpringImplicitNamingStrategy.class.getName());
-		}));
-	}
-
-	@Test
-	public void hibernate5CustomNamingStrategies() {
-		this.contextRunner.withPropertyValues(
-				"spring.jpa.hibernate.naming.implicit-strategy:com.example.Implicit",
-				"spring.jpa.hibernate.naming.physical-strategy:com.example.Physical")
-				.run(assertJpaProperties((properties) -> {
-					Map<String, Object> hibernateProperties = properties
-							.getHibernateProperties(new HibernateSettings());
-					assertThat(hibernateProperties).contains(
-							entry("hibernate.implicit_naming_strategy",
-									"com.example.Implicit"),
-							entry("hibernate.physical_naming_strategy",
-									"com.example.Physical"));
-					assertThat(hibernateProperties)
-							.doesNotContainKeys("hibernate.ejb.naming_strategy");
-				}));
-	}
-
-	@Test
-	public void hibernate5CustomNamingStrategiesViaJpaProperties() {
-		this.contextRunner.withPropertyValues(
-				"spring.jpa.properties.hibernate.implicit_naming_strategy:com.example.Implicit",
-				"spring.jpa.properties.hibernate.physical_naming_strategy:com.example.Physical")
-				.run(assertJpaProperties((properties) -> {
-					Map<String, Object> hibernateProperties = properties
-							.getHibernateProperties(new HibernateSettings());
-					// You can override them as we don't provide any default
-					assertThat(hibernateProperties).contains(
-							entry("hibernate.implicit_naming_strategy",
-									"com.example.Implicit"),
-							entry("hibernate.physical_naming_strategy",
-									"com.example.Physical"));
-					assertThat(hibernateProperties)
-							.doesNotContainKeys("hibernate.ejb.naming_strategy");
-				}));
-	}
-
-	@Test
-	public void useNewIdGeneratorMappingsDefault() {
-		this.contextRunner.run(assertJpaProperties((properties) -> {
-			Map<String, Object> hibernateProperties = properties
-					.getHibernateProperties(new HibernateSettings());
-			assertThat(hibernateProperties).containsEntry(
-					AvailableSettings.USE_NEW_ID_GENERATOR_MAPPINGS, "true");
-		}));
-	}
-
-	@Test
-	public void useNewIdGeneratorMappingsFalse() {
-		this.contextRunner
-				.withPropertyValues(
-						"spring.jpa.hibernate.use-new-id-generator-mappings:false")
-				.run(assertJpaProperties((properties) -> {
-					Map<String, Object> hibernateProperties = properties
-							.getHibernateProperties(new HibernateSettings());
-					assertThat(hibernateProperties).containsEntry(
-							AvailableSettings.USE_NEW_ID_GENERATOR_MAPPINGS, "false");
-				}));
-	}
 
 	@Test
 	public void determineDatabaseNoCheckIfDatabaseIsSet() {
@@ -174,30 +81,6 @@ public class JpaPropertiesTests {
 							.determineDatabase(mockDataSource("jdbc:h2:mem:testdb"));
 					assertThat(database).isEqualTo(Database.MYSQL);
 				}));
-	}
-
-	@Test
-	public void defaultDdlAutoIsNotInvokedIfPropertyIsSet() {
-		this.contextRunner.withPropertyValues("spring.jpa.hibernate.ddl-auto=validate")
-				.run(assertDefaultDdlAutoNotInvoked("validate"));
-	}
-
-	@Test
-	public void defaultDdlAutoIsNotInvokedIfHibernateSpecificPropertyIsSet() {
-		this.contextRunner
-				.withPropertyValues("spring.jpa.properties.hibernate.hbm2ddl.auto=create")
-				.run(assertDefaultDdlAutoNotInvoked("create"));
-	}
-
-	private ContextConsumer<AssertableApplicationContext> assertDefaultDdlAutoNotInvoked(
-			String expectedDdlAuto) {
-		return assertJpaProperties((properties) -> {
-			Map<String, Object> hibernateProperties = properties.getHibernateProperties(
-					new HibernateSettings().ddlAuto(this.ddlAutoSupplier));
-			assertThat(hibernateProperties).containsEntry("hibernate.hbm2ddl.auto",
-					expectedDdlAuto);
-			verify(this.ddlAutoSupplier, never()).get();
-		});
 	}
 
 	@Test
